Move components into sub-folders
